+++ tests/write.js
@@ -12,10 +12,10 @@ var data = [
 var ws_name = "SheetJS";
 
 var wscols = [
-	{wch:6},
-	{wch:7},
+	{wch:6}, // "characters"
+	{wpx:50}, // "pixels"
 	{wch:10},
-	{wch:20}
+	{wpx:125}
 ];
 
 
@@ -28,52 +28,11 @@ console.log("Columns :"); for(i=0; i!=wscols.length;++i) console.log(wscols[i]);
 /* require XLSX */
 if(typeof XLSX === "undefined") { try { XLSX = require('./'); } catch(e) { XLSX = require('../'); } }
 
-/* dummy workbook constructor */
-function Workbook() {
-	if(!(this instanceof Workbook)) return new Workbook();
-	this.SheetNames = [];
-	this.Sheets = {};
-}
-var wb = new Workbook();
-
-
-function datenum(v/*:Date*/, date1904/*:?boolean*/)/*:number*/ {
-	var epoch = v.getTime();
-	if(date1904) epoch += 1462*24*60*60*1000;
-	return (epoch + 2209161600000) / (24 * 60 * 60 * 1000);
-}
+/* blank workbook constructor */
+var wb = { SheetNames: [], Sheets: {} };
 
 /* convert an array of arrays in JS to a CSF spreadsheet */
-function sheet_from_array_of_arrays(data, opts) {
-	var ws = {};
-	var range = {s: {c:10000000, r:10000000}, e: {c:0, r:0 }};
-	for(var R = 0; R != data.length; ++R) {
-		for(var C = 0; C != data[R].length; ++C) {
-			if(range.s.r > R) range.s.r = R;
-			if(range.s.c > C) range.s.c = C;
-			if(range.e.r < R) range.e.r = R;
-			if(range.e.c < C) range.e.c = C;
-			var cell = {v: data[R][C] };
-			if(cell.v == null) continue;
-			var cell_ref = XLSX.utils.encode_cell({c:C,r:R});
-
-			/* TEST: proper cell types and value handling */
-			if(typeof cell.v === 'number') cell.t = 'n';
-			else if(typeof cell.v === 'boolean') cell.t = 'b';
-			else if(cell.v instanceof Date) {
-				cell.t = 'n'; cell.z = XLSX.SSF._table[14];
-				cell.v = datenum(cell.v);
-			}
-			else cell.t = 's';
-			ws[cell_ref] = cell;
-		}
-	}
-
-	/* TEST: proper range */
-	if(range.s.c < 10000000) ws['!ref'] = XLSX.utils.encode_range(range);
-	return ws;
-}
-var ws = sheet_from_array_of_arrays(data);
+var ws = XLSX.utils.aoa_to_sheet(data, {cellDates:true});
 
 /* TEST: add worksheet to workbook */
 wb.SheetNames.push(ws_name);
@@ -96,8 +55,18 @@ ws["!ref"] = "A1:E4";
 /* TEST: column widths */
 ws['!cols'] = wscols;
 
+/* TEST: hyperlink note: Excel does not automatically style hyperlinks */
+ws['A3'].l = { Target: "http://sheetjs.com", Tooltip: "Visit us <SheetJS.com!>" };
+
+/* TEST: built-in format */
+//ws['A1'].z = "0%"; wb.SSF[9] = "0%"; // Format Code 9
+
+/* TEST: custom format */
+//ws['B2'].z = "0.0"; wb.SSF[60] = "0.0"; // Custom
 console.log("JSON Data: "); console.log(XLSX.utils.sheet_to_json(ws, {header:1}));
 
+console.log("Worksheet Model:")
+console.log(ws);
 
 /* write file */
 XLSX.writeFile(wb, 'sheetjs.xlsx', {bookSST:true});